requirements.txt
Django==1.11.4
django-model-utils==3.0.0
olefile==0.44
Pillow==4.2.1
psycopg2==2.7.3
pytz==2017.2
wheel==0.24.0
